Fold cohesive signature clusters into typed structs and split git model by concern
